# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "REST API для системы осмотра жилых помещений ЖКХ.\nСистема предназначена для координации работы инспекторов,\nуправления заданиями на осмотр зданий и формирования отчётности.",
        "title": "JKH Inspection API",
        "contact": {
            "name": "Техническая поддержка",
            "email": "support@jkh.local"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
...
        "/inspector/tasks/{id}/act": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Скачивание PDF-акта осмотра здания",
                "produces": [
                    "application/pdf"
                ],
                "tags": [
                    "Инспектор"
                ],
                "summary": "Скачать акт осмотра",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID задания",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "PDF файл акта осмотра",
                        "schema": {
                            "type": "file"
                        }
                    },
                    "400": {
                        "description": "Неверный ID",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "401": {
                        "description": "Не авторизован",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Акт осмотра не найден",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Ошибка генерации акта",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/inspector/tasks/{id}/results": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Возвращает все результаты осмотра для конкретного задания",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Инспектор"
                ],
                "summary": "Получить результаты осмотра",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID задания",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Список результатов осмотра",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.InspectionResultResponse"
                            }
                        }
                    },
                    "400": {
                        "description": "Неверный ID",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "401": {
                        "description": "Не авторизован",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Задание не найдено",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Создание или обновление результата проверки элемента чек-листа",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Инспектор"
                ],
                "summary": "Создать/обновить результат осмотра",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID задания",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Результат осмотра элемента",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.CreateInspectionResultRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Результат сохранен",
                        "schema": {
                            "$ref": "#/definitions/models.InspectionResultResponse"
                        }
                    },
                    "400": {
                        "description": "Неверный запрос или задание не в работе",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "401": {
                        "description": "Не авторизован",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Задание не найдено",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/inspector/tasks/{id}/results/{element_id}": {
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Удаление результата осмотра конкретного элемента",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Инспектор"
                ],
                "summary": "Удалить результат осмотра",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID задания",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "ID элемента чек-листа",
                        "name": "element_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Результат успешно удален"
                    },
                    "400": {
                        "description": "Неверный ID",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "401": {
                        "description": "Не авторизован",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Результат не найден",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "models.CreateInspectionResultRequest": {
            "type": "object",
            "required": [
                "checklist_element_id",
                "condition_status"
            ],
            "properties": {
                "checklist_element_id": {
                    "description": "ID элемента чек-листа (из ChecklistElement).",
                    "type": "integer",
                    "minimum": 1
                },
                "comment": {
                    "description": "Комментарий инспектора (опционально).",
                    "type": "string"
                },
                "condition_status": {
                    "description": "Статус состояния элемента.\nДопустимые значения: \"Исправное\", \"Удовлетворительное\", \"Неудовлетворительное\", \"Аварийное\"",
                    "type": "string",
                    "enum": [
                        "Исправное",
                        "Удовлетворительное",
                        "Неудовлетворительное",
                        "Аварийное"
                    ]
                }
            }
        },
...
        "models.InspectionResultResponse": {
            "type": "object",
            "properties": {
                "checklist_element_id": {
                    "type": "integer"
                },
                "comment": {
                    "type": "string"
                },
                "condition_status": {
                    "description": "Результат проверки",
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "element_category": {
                    "description": "Категория",
                    "type": "string"
                },
                "element_name": {
                    "description": "Информация об элементе",
                    "type": "string"
                },
                "order_index": {
                    "description": "Порядок в чек-листе",
                    "type": "integer"
                },
                "task_id": {
                    "type": "integer"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Введите JWT токен в формате: Bearer {token}",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}